requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
openai>=1.0.0
pydantic>=2.0
python-dotenv>=1.0.0
streamlit>=1.30.0
fastapi>=0.110.0
uvicorn>=0.27.0
langgraph>=0.0.40
google-genai>=1.0.0
python-multipart>=0.0.5